Add null check to In-App Message Intent handling
It was assumed that the Intent from the current Activity will never be `null`.

In practise, we have observed the following crash:

```
Caused by: java.lang.NullPointerException: Attempt to invoke virtual method 'int android.content.Intent.getIntExtra(java.lang.String, int)' on a null object reference
        at com.optimove.android.optimobile.InAppMessagePresenter.activityAvailable(InAppMessagePresenter.java:54)
        at com.optimove.android.optimobile.AppStateWatcher.onActivityResumed(AppStateWatcher.java:82)
        at android.app.Application.dispatchActivityResumed(Application.java:416)
        at android.app.Activity.dispatchActivityResumed(Activity.java:1397)
        at android.app.Activity.onResume(Activity.java:1950)
        at androidx.fragment.app.FragmentActivity.onResume(FragmentActivity.java:434)
```

It is not clearly documented if an Activity's current Intent can be null, but nothing prevents it being set to null with Activity#setIntent(null).

So, we add a guard that prevents trying to unpack the IAM tickle ID if the Intent of the current Activity was null.

Note we don't abort the foreground handler so that existing behaviour of IAM presentation remains unchanged. Early abort could potentially leave an inconsistent view state for the presentation queue.